Extract `src/compile/scale/range`
